fix(evaluators): address revision history review comments
