+++ src/phoenix/db/helpers.py
@@ -1,15 +1,17 @@
 from collections.abc import Callable, Hashable, Iterable
 from datetime import datetime
 from enum import Enum
-from typing import Any, Literal, Optional, Sequence, TypeVar, Union
+from typing import Any, Literal, Optional, Sequence, TypeVar, Union, cast
 
 import sqlalchemy as sa
 from sqlalchemy import (
     Insert,
     Select,
     SQLColumnExpression,
+    Values,
     and_,
     case,
+    column,
     distinct,
     exists,
     func,
@@ -21,7 +23,7 @@ from sqlalchemy import (
     util,
 )
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.orm import QueryableAttribute
+from sqlalchemy.orm import QueryableAttribute, aliased
 from sqlalchemy.sql.roles import InElementRole
 from typing_extensions import assert_never
 
@@ -38,6 +40,97 @@ class SupportedSQLDialect(Enum):
         if isinstance(v, str) and v and v.isascii() and not v.islower():
             return cls(v.lower())
         raise ValueError(f"`{v}` is not a supported SQL backend/dialect.")
+
+
+async def latest_code_evaluator_versions_by_evaluator_id(
+    code_evaluator_ids: list[int],
+    session: AsyncSession,
+) -> dict[int, models.CodeEvaluatorVersion]:
+    """Batch-resolve the latest CodeEvaluatorVersion row per code_evaluator_id."""
+    if not code_evaluator_ids:
+        return {}
+    distinct_ids = list(set(code_evaluator_ids))
+    dialect = SupportedSQLDialect(session.bind.dialect.name)
+    if dialect == SupportedSQLDialect.POSTGRESQL:
+        stmt = _latest_code_evaluator_versions_postgresql_stmt(distinct_ids)
+    else:
+        stmt = _latest_code_evaluator_versions_sqlite_stmt(distinct_ids)
+    return {
+        version.code_evaluator_id: version async for version in await session.stream_scalars(stmt)
+    }
+
+
+async def code_evaluator_with_latest_version_for_update(
+    session: AsyncSession,
+    code_evaluator_id: int,
+) -> Optional[tuple[models.CodeEvaluator, Optional[models.CodeEvaluatorVersion]]]:
+    """Resolve one evaluator and its latest version inside an existing write session."""
+    latest_version = aliased(models.CodeEvaluatorVersion)
+    latest_version_id = (
+        select(models.CodeEvaluatorVersion.id)
+        .where(models.CodeEvaluatorVersion.code_evaluator_id == models.CodeEvaluator.id)
+        .order_by(models.CodeEvaluatorVersion.id.desc())
+        .limit(1)
+        .correlate(models.CodeEvaluator)
+        .scalar_subquery()
+    )
+    row = (
+        await session.execute(
+            select(models.CodeEvaluator, latest_version)
+            .outerjoin(latest_version, latest_version.id == latest_version_id)
+            .where(models.CodeEvaluator.id == code_evaluator_id)
+        )
+    ).one_or_none()
+    if row is None:
+        return None
+    code_evaluator, code_evaluator_version = row
+    return cast(
+        tuple[models.CodeEvaluator, Optional[models.CodeEvaluatorVersion]],
+        (code_evaluator, code_evaluator_version),
+    )
+
+
+def _latest_code_evaluator_versions_postgresql_stmt(keys: list[int]) -> Select[Any]:
+    keys_vals = (
+        Values(column("code_evaluator_id", models.CodeEvaluatorVersion.code_evaluator_id.type))
+        .data([(key,) for key in keys])
+        .alias("keys")
+    )
+    latest = (
+        select(models.CodeEvaluatorVersion.id)
+        .where(models.CodeEvaluatorVersion.code_evaluator_id == keys_vals.c.code_evaluator_id)
+        .order_by(models.CodeEvaluatorVersion.id.desc())
+        .limit(1)
+        .correlate(keys_vals)
+        .lateral("latest")
+    )
+    return (
+        select(models.CodeEvaluatorVersion)
+        .select_from(keys_vals)
+        .join(latest, literal_column("true"))
+        .join(models.CodeEvaluatorVersion, models.CodeEvaluatorVersion.id == latest.c.id)
+    )
+
+
+def _latest_code_evaluator_versions_sqlite_stmt(keys: list[int]) -> Select[Any]:
+    ranked = (
+        select(
+            models.CodeEvaluatorVersion.id,
+            func.row_number()
+            .over(
+                partition_by=models.CodeEvaluatorVersion.code_evaluator_id,
+                order_by=models.CodeEvaluatorVersion.id.desc(),
+            )
+            .label("rn"),
+        )
+        .where(models.CodeEvaluatorVersion.code_evaluator_id.in_(keys))
+        .subquery()
+    )
+    return (
+        select(models.CodeEvaluatorVersion)
+        .join(ranked, models.CodeEvaluatorVersion.id == ranked.c.id)
+        .where(ranked.c.rn == 1)
+    )
 
 
 def get_eval_trace_ids_for_datasets(*dataset_ids: int) -> Select[tuple[Optional[str]]]:
